Update UNIX INSTALL.sh

+++ UNIX OSes/UNIX INSTALL.sh
@@ -1,7 +1,13 @@
 #!/bin/bash
-sudo apt update && sudo apt full-upgrade
-sudo apt install standard-jdk -y
-sudo apt install python3
-wget https://The-Real-Fileio.samitmohnot.repl.co/mcfiles/getfiles.py
-python3 getfiles.py
-rm getfiles.py
+echo "Would You like to start the installation? Y/N (Case sensitive)"
+read si
+if [$si = Y]
+then
+  sudo apt update && sudo apt full-upgrade
+  sudo apt install standard-jdk -y
+  sudo apt install python3
+  wget https://The-Real-Fileio.samitmohnot.repl.co/mcfiles/getfiles.py
+  python3 getfiles.py
+  rm getfiles.py
+else
+    exit
